requirements.txt
# Stock & Crypto Ticker Plugin Dependencies
# Core dependencies for API requests and image processing

requests>=2.25.0
Pillow>=8.0.0
urllib3>=1.26.0
numpy>=1.20.0

# Optional dependencies for enhanced functionality
# pandas>=1.3.0  # For data analysis (if needed)
